Commonize VectorStore
