+++ chat_hatenablog/__init__.py
@@ -1,1 +1,3 @@
 __version__ = '0.0.1'
+
+from .vector_store import VectorStore
